- Unit test for SimpleIndexesSelector were created - SecondaryIndexProvidersInserter was created - Unit test for SecondaryIndexProvidersInserter were created

+++ Xtensive.Storage/Xtensive.Storage.Rse/Optimization/IndexSelection/ICostEvaluator.cs
@@ -12,6 +12,6 @@ namespace Xtensive.Storage.Rse.Optimization.IndexSelection
 {
   internal interface ICostEvaluator
   {
-    float Evaluate(IndexInfo indexInfo, RangeSet<Entire<Tuple>> rangeSet);
+    double Evaluate(IndexInfo indexInfo, RangeSet<Entire<Tuple>> rangeSet);
   }
 }
